Fix issue #4 Population initialization
Previously population initialization was always done with a gaussian distribution (?? don't know why) and the keep_bounds function was used to ensure the population was intialized between corrent boundaries. That previous approach limited the initial search space for other problems that had boundaries beyond that range. Thus, the code for population initialization has been updated and now it is done in [lower_bound, upper_bound] for each feature